+++ react-app/src/Assignment-3/3.css
@@ -1,109 +1,96 @@
-body {
-    display: flex;
-    justify-content: space-evenly;
-   }
 button {
-     padding: 10px;
-     cursor: pointer;
+  padding: 10px;
+  cursor: pointer;
 }
 code {
-    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-      monospace;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
+    monospace;
 }
-.auth{
-    display: flex;
-    justify-content: space-evenly;
+.auth {
+  display: flex;
+  justify-content: space-evenly;
 }
 .toggle-button {
-    margin: 1rem;
-    display: flex;
-    justify-content:center;
-    align-items: center;
+  margin: 1rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 body.light {
-    background-color: #ffffff;
-  }
-  
+  background-color: #ffffff;
+}
+
 body.dark {
-    background-color: #333333;
-    color: #ffffff;
-  }
-.link3{
-    display: flex;
-    justify-content: space-between;
-    border: 5rem;
-    border-color: blue;
-    margin-top: 25px;
-  
-  }
+  background-color: #333333;
+  color: #ffffff;
+}
+.link3 {
+  display: flex;
+  justify-content: space-between;
+  border: 5rem;
+  border-color: blue;
+  margin-top: 25px;
+}
 .link3 a {
-    text-decoration: none;
-    color: #ffffff;
-    padding: 15px 40px;
-    border-radius: 5px;
-    background-color: #1c3964;
-  
+  text-decoration: none;
+  color: #ffffff;
+  padding: 15px 40px;
+  border-radius: 5px;
+  background-color: #1c3964;
 }
-.link4{
-    display: flex;
-    justify-content: space-between;
-    border: 5rem;
-    border-color: blue;
-    margin-top: 25px;
-  
-  }
+.link4 {
+  display: flex;
+  justify-content: space-between;
+  border: 5rem;
+  border-color: blue;
+  margin-top: 25px;
+}
 .link4 a {
-    text-decoration: none;
-    color: #ffffff;
-    padding: 15px 40px;
-    border-radius: 5px;
-    background-color: #1c3964;
-  
+  text-decoration: none;
+  color: #ffffff;
+  padding: 15px 40px;
+  border-radius: 5px;
+  background-color: #1c3964;
 }
-.link5{
-    display: flex;
-    justify-content:space-between;
-    border: 5rem;
-    border-color: blue;
-    margin-top: 25px;
-  
+.link5 {
+  display: flex;
+  justify-content: space-between;
+  border: 5rem;
+  border-color: blue;
+  margin-top: 25px;
 }
 .link5 a {
-    text-decoration: none;
-    color: #ffffff;
-    padding: 15px 40px;
-    border-radius: 5px;
-    background-color: #1c3964;
-  
+  text-decoration: none;
+  color: #ffffff;
+  padding: 15px 40px;
+  border-radius: 5px;
+  background-color: #1c3964;
 }
 .dashboard a {
-    text-decoration: none;
-    color: #ffffff;
-    padding: 15px 40px;
-    border-radius: 5px;
-    background-color: #1c3964;
-  
+  text-decoration: none;
+  color: #ffffff;
+  padding: 15px 40px;
+  border-radius: 5px;
+  background-color: #1c3964;
 }
-.dashboard{
-    display: flex;
-    justify-content: space-between;
-    border: 5rem;
-    border-color: blue;
-    margin-top: 25px;
+.dashboard {
+  display: flex;
+  justify-content: space-between;
+  border: 5rem;
+  border-color: blue;
+  margin-top: 25px;
 }
-.link6{
-    display: flex;
-    justify-content:space-between;
-    border: 5rem;
-    border-color: blue;
-    margin-top: 25px;
-  
+.link6 {
+  display: flex;
+  justify-content: space-between;
+  border: 5rem;
+  border-color: blue;
+  margin-top: 25px;
 }
 .link6 a {
-    text-decoration: none;
-    color: #ffffff;
-    padding: 15px 40px;
-    border-radius: 5px;
-    background-color: #1c3964;
-  
+  text-decoration: none;
+  color: #ffffff;
+  padding: 15px 40px;
+  border-radius: 5px;
+  background-color: #1c3964;
 }
