+++ react-app/src/Assignment-3/Assignment3.jsx
@@ -2,9 +2,7 @@ import {
   AuthProvider,
   PreferencesProvider,
   Routing,
-  NestedRoute,
-  Detail,
-  ProtectedAuthentication,
+  AllRoute,
 } from "./index.jsx";
 import { NavLink } from "react-router-dom";
 import "./3.css";
@@ -35,44 +33,9 @@ const Assignment3 = () => {
         </div>
       </div>
 
-      <div className="question-4">
-        <h2>Question-4</h2>
-        <ProtectedAuthentication />
-        <div className="link4">
-          <NavLink exact to="/">
-            Home
-          </NavLink>
-          <NavLink exact to="/about">
-            About Us
-          </NavLink>
-          <NavLink exact to="/login">
-            Login
-          </NavLink>
-        </div>
-      </div>
-
-      <div className="question-5">
-        <h2>Question-5</h2>
-        <NestedRoute />
-        <div className="link5">
-          <NavLink exact to="/">
-            Home
-          </NavLink>
-          <NavLink exact to="/about">
-            About Us
-          </NavLink>
-          <NavLink exact to="/login">
-            Login
-          </NavLink>
-          <NavLink exact to="/dashboard">
-            DashBoard
-          </NavLink>
-        </div>
-      </div>
-
-      <div className="question-6">
-        <h2>Question-6</h2>
-        <Detail />
+      <div className="question-456">
+        <h2>Question-456</h2>
+        <AllRoute />
         <div className="link6">
           <NavLink exact to="/">
             Home
