+++ react-app/src/Assignment-3/Question456/Items.jsx
@@ -1,0 +1,8 @@
+const items = [
+  { id: 1, name: "Laptop ", price: 70000 },
+  { id: 2, name: "Mobile", price: 60000 },
+  { id: 3, name: "HeadPhones", price: 5000 },
+  { id: 4, name: "EarPods", price: 2000 },
+  { id: 5, name: "Charger", price: 1000 },
+];
+export { items };
