+++ react-app/src/Assignment-3/Question456/Products.jsx
@@ -1,15 +1,6 @@
 import React from "react";
 import { useParams, useNavigate } from "react-router-dom";
-
-const items = [
-  { id: 1, name: "Laptop ", price: 70000 },
-  { id: 2, name: "Mobile", price: 60000 },
-  { id: 3, name: "HeadPhones", price: 5000 },
-  { id: 4, name: "EarPods", price: 2000 },
-  { id: 5, name: "Charger", price: 1000 },
-  // Add more items as needed
-];
-
+import {items} from "./Items";
 const Products = () => {
   const navigate = useNavigate();
   const { id } = useParams();
@@ -36,4 +27,3 @@ const Products = () => {
 };
 
 export default Products;
-export { items };
