--- react-app/src/Assignment-3/Question5/About.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-
-const About=()=>{
-    return(
-      <h2>Welcome to the About Page</h2>
-    );
-}
-export default About;
