+++ react-app/src/Assignment-3/index.jsx
@@ -1,15 +1,6 @@
 import AuthProvider from "./Question1/Question-1";
 import PreferencesProvider from "./Question2/Question-2";
 import Routing from "./Question3/Question-3";
-import ProtectedAuthentication from "./Question4/Question-4";
-import NestedRoute from "./Question5/Question-5";
-import Detail from "./Question6/Question-6";
+import AllRoute from "./Question456/Routes";
 
-export {
-  AuthProvider,
-  PreferencesProvider,
-  Routing,
-  NestedRoute,
-  Detail,
-  ProtectedAuthentication,
-};
+export { AuthProvider, PreferencesProvider, Routing, AllRoute };
